requirements.txt
Django==3.1.1
django-cors-headers==3.10.1
django-filter==21.1
django-rest-knox==4.1.0
django-utils-six==2.0
djangorestframework==3.13.1
Pillow==8.4.0
python-decouple==3.5
